+++ R/add-proportional.R
@@ -45,13 +45,13 @@ ff_pie <- function(.type = "pie") {
 #'   add_donut()
 #'
 #' # for a `sum` provide `color` and `y`
-#' # `sum` of `power` in each `energy_type` category
+#' # `sum` of `energy` in each `energy_type` category
 #' energy %>%
-#'   tidyplot(y = power, color = energy_type) %>%
+#'   tidyplot(y = energy, color = energy_type) %>%
 #'   add_pie()
 #'
 #' energy %>%
-#'   tidyplot(y = power, color = energy_type) %>%
+#'   tidyplot(y = energy, color = energy_type) %>%
 #'   add_donut()
 #'
 #' @export
@@ -122,31 +122,31 @@ ff_barstack <- function(.position_fun) {
 #'   add_barstack_relative()
 #'
 #' # for a `sum` provide `color` and `y`
-#' # `sum` of `power` in each `energy_type` category
+#' # `sum` of `energy` in each `energy_type` category
 #' energy %>%
-#'   tidyplot(y = power, color = energy_type) %>%
+#'   tidyplot(y = energy, color = energy_type) %>%
 #'   add_barstack_absolute()
 #'
 #' energy %>%
-#'   tidyplot(y = power, color = energy_type) %>%
+#'   tidyplot(y = energy, color = energy_type) %>%
 #'   add_barstack_relative()
 #'
 #' # Include variable on second axis
 #' energy %>%
-#'   tidyplot(x = year, y = power, color = energy_type) %>%
+#'   tidyplot(x = year, y = energy, color = energy_type) %>%
 #'   add_barstack_absolute()
 #'
 #' energy %>%
-#'   tidyplot(x = year, y = power, color = energy_type) %>%
+#'   tidyplot(x = year, y = energy, color = energy_type) %>%
 #'   add_barstack_relative()
 #'
 #' # Flip x and y-axis
 #' energy %>%
-#'   tidyplot(x = power, y = year, color = energy_type) %>%
+#'   tidyplot(x = energy, y = year, color = energy_type) %>%
 #'   add_barstack_absolute(orientation = "y")
 #'
 #' energy %>%
-#'   tidyplot(x = power, y = year, color = energy_type) %>%
+#'   tidyplot(x = energy, y = year, color = energy_type) %>%
 #'   add_barstack_relative(orientation = "y")
 #'
 #' @export
@@ -275,22 +275,22 @@ ff_areastack <- function(.position_fun) {
 #'   add_areastack_relative()
 #'
 #' # for a `sum` provide `x`, `y` and `color`
-#' # `sum` of `power` in each `energy_type` category
+#' # `sum` of `energy` in each `energy_type` category
 #' energy %>%
-#'   tidyplot(x = year, y = power, color = energy_type) %>%
+#'   tidyplot(x = year, y = energy, color = energy_type) %>%
 #'   add_areastack_absolute()
 #'
 #' energy %>%
-#'   tidyplot(x = year, y = power, color = energy_type) %>%
+#'   tidyplot(x = year, y = energy, color = energy_type) %>%
 #'   add_areastack_relative()
 #'
 #' # Flip x and y-axis
 #' energy %>%
-#'   tidyplot(x = power, y = year, color = energy_type) %>%
+#'   tidyplot(x = energy, y = year, color = energy_type) %>%
 #'   add_areastack_absolute(orientation = "y")
 #'
 #' energy %>%
-#'   tidyplot(x = power, y = year, color = energy_type) %>%
+#'   tidyplot(x = energy, y = year, color = energy_type) %>%
 #'   add_areastack_relative(orientation = "y")
 #'
 #' @export