+++ data-raw/energy.R
@@ -1,10 +1,29 @@
 ## code to prepare `energy` dataset goes here
 
 library(tidyverse)
+library(tidyplots)
+
+# corrected version
+read_single_file <- function(x) {
+  read_csv(x, comment = "#") %>%
+    pivot_longer(-Year, names_to = "energy_source", values_to = "energy") %>%
+    rename(year = Year)
+}
 
 energy <-
-  read_csv("data-raw/energy_total_year.csv") %>%
-  pivot_longer(-year, names_to = "energy_source", values_to = "power") %>%
+  list.files(path = "data-raw/energy_charts_download",
+             pattern = "in_20",
+             full.names = TRUE
+             ) %>%
+  map(read_single_file) %>%
+  bind_rows() %>%
+  mutate(
+    energy_source = str_replace_all(energy_source, "Fossil.*gas", "Fossil gas"),
+    energy_source = str_replace_all(energy_source, "Solar.*", "Solar"),
+    energy_source = str_replace_all(energy_source, "Hydro.*", "Hydro"),
+    energy_source = str_replace_all(energy_source, "Waste.*", "Waste")
+  ) %>%
+  tidyr::complete(year, energy_source, fill = list(energy = 0)) %>%
   mutate(energy_type = case_when(
     str_detect(energy_source, "Geo|Hydro|Wind|Bio|Solar") ~ "Renewable",
     str_detect(energy_source, "Nuclear") ~ "Nuclear",
@@ -12,11 +31,12 @@ energy <-
     .default = "Other"
   )) %>%
   mutate(
-    power_unit = "GW",
+    energy_unit = "TWh",
     energy_source = factor(energy_source),
     energy_type = factor(energy_type)
-    ) %>%
-  tidyr::replace_na(list(power = 0)) %>%
-  relocate(year, energy_source, energy_type, power, power_unit)
+  ) %>%
+  relocate(year, energy_source, energy_type, energy, energy_unit)
 
 usethis::use_data(energy, overwrite = TRUE)
+
+# write_csv(energy, "electricity-generation-in-Germany.csv")