+++ NEWS.md
@@ -1,3 +1,5 @@
+# tidyplots (development version)
+
 # tidyplots 0.2.0
 
 ## Bug fixes