+++ NEWS.md
@@ -3,6 +3,7 @@
 * New color scheme `colors_discrete_rainbow` (@electrolars, #35).
 * The function `save_plot()` gains the new argument `view_plot` 
 to control whether to view the plot on screen after saving (#40).
+* Fixed a bug when combining counts with axis limits (#41).
 
 # tidyplots 0.2.0