+++ NEWS.md
@@ -1,4 +1,4 @@
-# tidyplots (development version)
+# tidyplots 0.2.1
 
 ## Breaking changes